refactor codes
