+++ UnitTestProject1/RandomDataGenerator.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace LN.PM.FirmManager.UIAutomation.Specs.Utilities
+namespace AnimalDownload
 {
     public static class RandomDataGenerator
     {
